correction + render image

+++ heightfield.h
@@ -9,9 +9,9 @@ class HeightField : public ScalarField2
 {
 public:
     HeightField();
-    HeightField(Vector2 a, Vector2 b, int ii, int jj, float defaut=0.0);
+    HeightField(Vector2 a, Vector2 b, int ii, int jj, double defaut=0.0);
     Vector3 normal(int i, int j);
-    void CalcUV(const Vector2 & p , int & xi, int & yi, float & u, float & v);
+    void CalcUV(const Vector2 & p , int & xi, int & yi, double & u, double & v);
     void Bilineaire(const Vector2 & p, double & res);
     void Barycentrique(const Vector2 & p, double & res);
